Restrict ability to delete items

+++ src/resolvers/Mutation.js
@@ -46,9 +46,15 @@ const Mutations = {
     async deleteItem(parent, args, ctx, info) {
         const where =  { id: args.id };
         // Find the item
-        const item = await ctx.db.query.item({ where }, `{ id title }`);
+        const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
         // Check if they own that item or have permissions
-        // TODO
+        const ownsItem = item.userid === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission => {
+            ['ADMIN', 'ITEMDELETE'].includes(permission)});
+
+        if (!ownsItem && hasPermissions) {
+            throw new Error("You don't have permission to do that!")
+        }
         // if so, delete it
         return ctx.db.mutation.deleteItem({ where }, info)
     },
